[DEV] sorting node function
