+++ front/src/components/FlowCanvas.css
@@ -1,0 +1,4 @@
+.wrapper {
+  width: 100vw;
+  height: 100vh;
+}
